Fix compiler warning (unportable sprintf usage)

+++ src/backend/storage/freespace/freespace.c
@@ -382,7 +382,8 @@ fsm_space_needed_to_cat(Size needed)
 
 	/* Can't ask for more space than the highest category represents */
 	if (needed > MaxFSMRequestSize)
-		elog(ERROR, "invalid FSM request size %d", needed);
+			elog(ERROR, "invalid FSM request size %lu",
+				 (unsigned long) needed);
 
 	if (needed == 0)
 		return 1;
